World Boundaries
Added soft world boundaries. What this means is that if the player goes
off the screen, a warning will pop up and after 10 seconds they will be
pushed back towards the game world. It is possible to overpower this and
just keep flying out of bounds, so we might want to add a hard reset if
the player goes too far.